Fix warning C4018: '<=': signed/unsigned mismatch.

+++ src/color.c
@@ -516,7 +516,7 @@ void C_puts_long_line (const char *start, size_t indent)
 
       if (!p)
           p = strchr (c+1, '\0');
-      if ( left < 2 || (left <= p - c) )
+      if (left < 2 || (left <= (size_t)(p - c)))
       {
         C_printf ("\n%*c", (int)indent, ' ');
         left = c_screen_width - indent;
